requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
haystack-ai>=2.0.0
